+++ backend/config/views.py
@@ -19,7 +19,7 @@ from rest_framework.decorators import authentication_classes
 from rest_framework.decorators import permission_classes
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.settings import api_settings
-
+from graphene_file_upload.django import FileUploadGraphQLView
 
 BASE_DIR = environ.Path(__file__) - 2  # settings.pyの2階層上のディレクトリ
 
@@ -96,20 +96,18 @@ class TwitterAccessToken(views.APIView):
         return Response(serializer.validated_data)
 
 
-class DRFAuthenticatedGraphQLView(GraphQLView):
+class DRFAuthenticatedGraphQLView(FileUploadGraphQLView):
     """
         GraphQLで、rest frameworkの認証を使うために、GraphQLViewをオーバーライド
         参照元: https://github.com/graphql-python/graphene/issues/249
+
+        継承関係: GraphQLView ← FileUploadGraphQLView ← DRFAuthenticatedGrqphQLView
+        FileUploadGraphQLViewで、parse_bodyがオーバーライドされ、fileを受け取れるようになっている
     """
-    def parse_body(self, request):
-        if isinstance(request, rest_framework.request.Request):
-            return request.data
-        return super(DRFAuthenticatedGraphQLView, self).parse_body(request)
 
     @classmethod
     def as_view(cls, *args, **kwargs):
-        view = super(DRFAuthenticatedGraphQLView, cls).as_view(*args, **kwargs)
-
+        view = super().as_view(*args, **kwargs)
         # if settings.DEBUG is False:
         #     view = permission_classes((IsAuthenticated,))(view)
         # else:
@@ -121,3 +119,8 @@ class DRFAuthenticatedGraphQLView(GraphQLView):
         view = api_view(["GET", "POST"])(view)
 
         return view
+
+    def parse_body(self, request):
+        # if isinstance(request, rest_framework.request.Request):
+        #     return request.data
+        return super(DRFAuthenticatedGraphQLView, self).parse_body(request)
